requirements.txt
# PIN PLAN - Financial planning MVP
# Use: pip install -r requirements.txt

streamlit>=1.28.0
dedalus-labs>=0.2.0
python-dotenv>=1.0.0
